requirements.txt
openai==1.35.7
pandas==2.2.2
requests==2.31.0
streamlit==1.34.0
python-dotenv==0.19.2